perf(pnl): avoid shifting FIFO lots
